Reorganise search results UI code
