+++ rca/static_src/sass/main.scss
@@ -83,7 +83,8 @@
 @import 'components/title-area';
 @import 'components/video-modal';
 
-@import '../javascript/programmes/components/ProgrammeTeaser';
+@import '../javascript/programmes/components/ProgrammesResults/ProgrammesResults';
+@import '../javascript/programmes/components/ProgrammesResults/ProgrammeTeaser';
 
 // Layout
 @import 'layout/footer';
